#121 structuring the modules/bank-feed
