+++ modules/bank-feed/data/bank-data.ts
@@ -1,0 +1,15 @@
+const BankData = {
+    id: '1234',
+    accountnumber: '1111222233334444',
+    ifsc: 'XYZN01234',
+    sortcode: '1234',
+    isbn: 'dkjdsj',
+    statement: [
+        {
+            date: '2222-22-22',
+            description: 'This is bank statement ',
+            amount: '50000',
+        },
+    ],
+};
+export default BankData;
